fix: repository URL sometimes formatted with `False`
The repository URL is sometimes incorrectly formatted with `False`
values instead of the correct environment variable. This happens because
the walrus operator assignment is not properly delimited. For instance,
the GitHub CI environment would incorrectly report the repo URL as:

```
WARNING  Repository URL differs from what would be set! Keeping existing value.
         Existing: https://github.com/maxrake/delme
         Proposed: False/False
         Use CLI to override: https://docs.phylum.io/docs/phylum_project_update
```

This change ensures the walrus operator assignment is parsed as intended.