refactor: trim log messages and remove walrus operators (#362)
This change seeks to clean up a few things that were discovered during
recent testing, and to do so before the next release. The use of walrus
operators (i.e., `:=`) has been removed in favor of more, but simpler,
code. Multiline log messages have been trimmed a bit after it was
discovered that lines with >71 characters wrap in constrained
environments where the WIDTH is fixed at 80 characters (e.g., GitHub
Actions). This is due to the space between the log level (e.g.,
`WARNING`) and the message being two characters instead of the assumed
one, allowing a max of 71 characters for each line of the message
instead of 72 as previously assumed. There is likely a better way to
programmatically account for this with something in the `textwrap`
module, but these small changes will do for now.