chore!: account for breaking Phylum CLI changes
The Phylum CLI v6.0.0 release introduces a number of breaking changes.
This PR seeks to account for those changes with these actions:

* Change `find-lockable-files` command to `find-dependency-files`
* Change `--lockfile-type` arguments to `--type`
* Change `lockfiles` key in `.phylum_project` file to `depfiles`
* Change `lockfiles` key in `phylum status --json` output to
  `dependency_files`
* Change custom `ci` extension to match the extension API updates
  * Change `analyze` packages type to `PackageWithOrigin`
* Rename `PackageDescriptor` dataclass to `Package`
  * Replace `lockfile` field with `origin`
* Ensure parsed dependency files include the "origin" path to that file
* Update tests

Closes #363

BREAKING CHANGE: Phylum CLI installs before v6.0.0-rc1 are no longer
supported. That release introduced a number of breaking changes which
are only functional with this release of the `phylum` package.