build: migrate from Python Semantic Release v7
This change mostly follows the Python Semantic Release (PSR) migration
guide found here:

https://python-semantic-release.readthedocs.io/en/latest/migrating_from_v7.html

Some basic changes include:

* Unpin `python-semantic-release` dependency and update `poetry.lock`
* Update the release and preview workflows to account for the changes
  * Use Poetry to upload the package to PyPI
  * Account for command and option changes
* Format the existing `CHANGELOG.md` file to be consistent throughout
* Update the release process documentation

A bigger change involves the use of Jinja templates to render the
CHANGELOG and release notes. A new directory, `templates`, was created
to supply the expected templates. To keep the old format, a snapshot of
the CHANGELOG was taken and gets used in full to represent all releases
that have come out before now. Macros and reusable templates are
employed to stay DRY.

The goal with the new templates was to re-create the same format used by
PSR prior to the v8 major release. That goal was mostly met, with one
notable diffence being the way breaking changes are represented. Instead
of separating the commit type from the list of breaking changes, any
commit that gets classified as "breaking" will be included in it's own
"Breaking" section, with the `BREAKING CHANGE` trailers forming
sub-bullets in that section now.

Closes #278