docs: add `--no-cache` option for `poetry update`
It was discovered through local use that using the cache may not always
result in generating the same results as the CI environment for the same
command issued at the same time. Specifying the `--no-cache` option
ensures that the results generated are the same, regardless of where the
command is run.