docs: restore integration documentation
This change restores the integration documentation that was previously
stored in this repository by copying it from the `documentation` repo.
That way, the code and docs will live together and be maintained more
easily. This is the first step in a two-step process. The second step is
to update the `documentation` repo to include these integration docs as
a git submodule, in the same way as is done for the `cli` repo. This
change is also acceptable because the `phylum-ci` repo is also public
and therefore available to view and suggest/submit changes directly from
the docs.phylum.io website.