docs: restore integration documentation (#395)
This change restores the integration documentation that was previously
stored in this repository by copying it from the `documentation` repo
(no additional changes were made). That way, the code and docs will live
together and be maintained more easily.

This is the first step in a three-step process. The second step is to
update the `documentation` repo to include these integration docs as a
git submodule, in the same way as is done for the `cli` repo. The third
step is to ensure releases of `phylum-ci` trigger a sub-module update in
the `documentation` repository.

This change is also acceptable because the `phylum-ci` repo is also
public and therefore available to view and suggest/submit changes
directly from the docs.phylum.io website.