feat: implement audit mode (#400)
This change introduces audit mode in the form of an `--audit` flag. It
also adds a `--skip-comments`/`-s` flag to disable posting
comments/notes on pull/merge requests. This flag is implicitly set when
audit mode is enabled. The documentation was updated to match.
Additionally, a change was made to use `inspect.cleandoc` instead of
`textwrap.dedent` everywhere. Using `cleandoc` is preferred since it
allows for removing the ugly line continuation character from multiline
strings.

Closes #396
Closes #311