refactor: adhere to new `ruff` preview rule SIM103
The latest release of `ruff` introduced the new linting rule
[SIM103: needless-bool](https://docs.astral.sh/ruff/rules/needless-bool)
which causes the code in question to fail when run in preview mode. This
change was made now to get ahead of the rule, for when it enforces the
implicit `else` cases by default. The change was made in a way that
seeks to maintain readability and intent. Type annotations were added to
the affected function as well.