feat: ensure Bitbucket repo URLs use HTTPS scheme (#406)
This change was made after discovering that Atlassian Compass components
recognize Bitbucket repository URLs only when they are `HTTPS:` and not
when they are `HTTP:`, even though that is how the URL is provided when
obtained by getting the `BITBUCKET_GIT_HTTP_ORIGIN` environment variable
value.

Additionally, the `is_token_set` function was refactored to adhere to
the new `ruff` preview rule `SIM103`. The latest release of `ruff`
introduced the new linting rule
[SIM103: needless-bool](https://docs.astral.sh/ruff/rules/needless-bool),
which causes the code in question to fail when run in preview mode. This
change was made now to get ahead of the rule, for when it enforces the
implicit `else` cases by default. The change was made in a way that
seeks to maintain readability and intent. Type annotations were added to
the affected function as well.