docs: recommend triggering scans for default branch (#407)
This change updates the docs for the integrations that allow for
triggering Phylum analysis upon pushes to the default branch: GitLab CI,
Azure Pipelines, and Bitbucket Pipelines (GitHub Actions currently only
runs in the context of `pull_request` webhook events).

The change is made to coincide with the upcoming platform feature
enabling default labels. It is meant to ensure the Phylum analysis
results for the branch associated with that label are current.