ci: account for `python-semantic-release` changes
The most recent weekly dependency bump caused the
`python-semantic-release` package to be upgraded from v9.4.2 --> v9.6.0,
which introduced breaking changes. With the irony noted, this PR seeks
to account for the changes with the following actions:

* Replace instances of `--prerelease` with `--as-prerelease`
  * This was the breaking change
* Make use of the new `compare_url` filter
  * Remove the custom macro that did the same thing
* Remove `.macros.j2` file since it only contains a constant used once

These changes were verified with local testing, using the `--noop`
option and comparing the output to the expected and existing output.