build: update `poetry` to latest version
Poetry is the workflow management tool used for this project and forms
the root of all other actions taken when working with this repository.
It is also used to manage dependencies and therefore should be treated
very carefully, with updates to newer versions taken deliberately.

This PR updates `poetry` to the latest version of v1.8.3 to account for
the [changes introduced](https://python-poetry.org/history) in both
v1.8.2 and v1.8.3 (full comparison
[here](https://github.com/python-poetry/poetry/compare/1.8.1...1.8.3)),
with these actions taken:

* Bump all instances of `poetry` to the new version
  * Installs in workflows
  * pre-commit hook revision
  * Dockerfiles
* Update the lockfile with the new version of `poetry`

None of the changes or new features in these new versions required any
updates to the use of `poetry` in this project. Additionally, there has
not been an update to the `poetry-core` build backend since the last
time a bump like this was made.