docs: update docs based on user feedback (#434)
This change seeks to improve the integration documentation by acting on
user feedback. In particular, the following suggestions were taken:

* Add detail on how to use audit mode with the `pre-commit` hook
* Make debug/verbose output the default in recommended CI configurations
* Be more explicit about the GitHub Action options
* The default action uses the `latest` tag from the `phylum-ci` Docker
image and should be used unless there is a good reason not to
* The alternative options have callouts and notes now making it more
clear that their configurations should only be used if the default
action does not suffice

A separate PR will be made in the `phylum-analyze-pr-action` repo, to
include the same changes here.