ci: ensure forced patch releases are possible
This change fixes a small logic error in the release process. The value used to compare whether a release needs to be forced as a patch release (due to no bump-worthy changes found by Python Semantic Release) was overwritten before it was saved off. That is no longer the case. The environment variable name wasalso changed to make the difference between what Python Semantic Release thinks the next release version should be and what the actual next release version will be.