feat: add support for Jenkins Pipelines
This change adds an integration for the Jenkins pipelines CI
environment. Specifically, it adds support for declarative pipeline
syntax in multibranch pipelines. Pipeline runs for both PRs and branches
are supported.

This integration detects and handles both branch pipelines and pull
request (PR) pipelines. When in the PR pipeline context, the results
are only written to logs and NOT as a comment on the PR. This is due to
the nature of the environment, where many different source repositories
are supported (e.g., GitHub, Bitbucket, GitLab, etc.)

The `documentation` repo will be updated separately to add this
integration to the summary overview page.

Closes #430