feat: add `ps` to Docker images
This change is based on a request from users of `podman`. Older versions
of that tool (prior to v4.6.0) require `ps` to exist on the running
container in order to use the `top` command while newer versions do not.

This package adds about 340kB to the compressed image size.

An image was built with these changes and provided for testing as the
`phylumio/phylum-ci:with_ps` tag on
[Docker Hub](https://hub.docker.com/r/phylumio/phylum-ci/tags). The user
in question has confirmed that the change works for them.