feat: allow groups to be specified without a project
This change allows users to specify a group without also specifying a
corresponding project. Neither the project nor the group needs to exist
beforehand. The group will now be created if it does not already exist.

Additionally, the values for project and group are now allowed to be
"mixed" between what is specified on the command line and anything found
in a `.phylum_project` file. This "mixing" was specifically denied
previously but allowed now with a warning to let users know that may be
an unexpected combination.

The group name is still optional and can be specified on the command
line. It can also be specified in the `.phylum_project` file. The value
specified with the option takes precedence when both are provided.

Other changes made include:

* Update documentation
  * Add an example of using the `--group` option to each integration
  * Format the Jenkins groovy samples to use 2-space tabs instead of 4
* Update `.phylum_project` file to use `depfiles` instead of `lockfiles`
* Rename CLI exit code enum entry for group or project already existing
  * From `PROJECT_ALREADY_EXISTS` to `ALREADY_EXISTS`
* Simplify some output messages
  * Use fewer words where possible
  * Be clear about when paid accounts are needed and provide info link