feat: add `git-lfs` to Docker images (#453)
This change adds support for [git large files
storage](https://git-lfs.com/) (lfs) by adding the `git-lfs` package to
both the default and `slim` Docker images. This change was prompted by a
user request and barely increases the uncompressed image size. The
feature will be useful for anyone making use of git lfs in the
repositories they wish to analyze with Phylum.