build: pin `python-semantic-release` to v9.8.3
This change pins the `python-semantic-release` (PSR) dependency to the
last known working version of v9.8.3. After that release, a number of
large refactors were made and introduced breaking changes, even though
they were only bugfix releases (I know, the irony). Specifically, the
`context` variable stopped being defined for the release notes template.

This variable is used in the `phylum-ci` project to create "compare"
links for making it easy to see what has changed from the previous
release. Instead of removing the compare link from the template, the PSR
dependency was pinned, with a note to remove the pin when PSR fixes the
regression.