build: update `python-semantic-release` pin (#463)
This change is an update to the one made in #454. Since then, the
`python-semantic-release` (PSR) dependency has responded to their
breaking change by re-defining the `context` variable for release notes
templates. However, they also say that the variable was not meant to be
exposed in the first place and that it will be removed again in the next
major version release. Therefore, the pin has been changed to a
caret-style one to allow for getting future releases that are at least
v9.8.7 but less than v10.

Discussions with the maintainer are ongoing, to request the features
relied upon with the current implementation are accounted for in the
next major version release.

References:
* https://github.com/python-semantic-release/python-semantic-release/pull/1005
* https://github.com/python-semantic-release/python-semantic-release/issues/984

Additionally, the release notes template was refactored a tiny bit to
account for new methods of doing the same thing, but with less code. The
changes were tested locally and found to produce the same output as the
previous template.