docs: add exit code information (#465)
This change documents the various exit codes in use for the `phylum-ci`
script entry point. The documentation for integrations that make use of
this script were updated to include a reference to the detail. The debug
output now includes a link to that same detail.

This change was generated from a user request. They were curious to know
why a failing status check was reported when a successful analysis
comment was added to their PR. They didn't understand what exit code 10
meant. This specific situation is already answered in a FAQ and now is
more likely to be seen since it is referenced in the exit code table.