fix: include `phylum-ci.exe` in release artifacts (#477)
This change corrects a bug in the release workflow where the
`phylum-ci.exe` artifact is not published to the newly created GitHub
release. The `python-semantic-release` tool has the following config, in
`pyproject.toml`, for publishing artifacts by pattern:

```
[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*", "build/phylum-ci.exe"]
```

The issue was that the `actions/upload-artifact` step used
`phylum-ci.exe` as the artifact name, which resulted in the following
directory structure for the corresponding `actions/download-artifact`
step:

`./phylum-ci.exe/phylum-ci.exe`

This clearly won't match the glob pattern of `build/phylum-ci.exe`. So,
this change is a simple one: change the upload artifact name to `build`
so that the downloaded artifact will exist at `./build/phylum-ci.exe`.