feat: add Windows standalone archive install option (#481)
This change adds another Windows standalone "installation" option, in
the form of a zip archive that contains all the files needed to run the
`phylum-ci.exe` binary...which is also included in the archive. This
option is not as convenient as the single file self-extracting
executable since it requires an additional unzipping step. It does have
the advantage of not getting flagged by antivirus.