ci: add waits for PyPI publications
This change accounts for the time it takes to publish a release on PyPI (or TestPyPI) and when that release is available for use. The test steps in CI would sometimes fail because they expected the release to be available immediately after publication. Now, the step will try for up to three minutes before failing the step, which appears to be plenty of time based on previous observations.