build: account for Python Semantic Release updates (#483)
Well, it happened again. Python Semantic Release (PSR) released a minor
version this week and it ended up having breaking changes. The irony is
too much and at this point all releases are examined carefully for any
breaking changes. This time, the v9.10.0 release of PSR included this
commit, which changed the long names of the types of angular commits as
they are referred to in a parsed release element:

https://github.com/python-semantic-release/python-semantic-release/pull/1045/commits/0766184986dd9625cd684c8602c35d147d407026

The effect would have been that, going forward, only commits with the
`documentation` type would have been picked up for inclusion. The
updates were accounted for in `templates/.changes.j2`.

The other big change, but not breaking, for the v9.10.0 release, was the
return of the placeholder style CHANGELOG insertion method. The
mechanics behind it have changed since the pre-v8 releases, but the end
result is the same. Since this method was and still is preferred,
changes were made to take advantage of it. Those changes followed the
guidance here:

https://python-semantic-release.readthedocs.io/en/stable/changelog_templates.html#incrementally-updating-changelog-template