build: update `poetry` to latest version
Poetry is the workflow management tool used for this project and forms
the root of all other actions taken when working with this repository.
It is also used to manage dependencies and therefore should be treated
very carefully, with updates to newer versions taken deliberately.

This PR updates `poetry` to the latest version of v1.8.4 to account for
the [changes introduced](https://python-poetry.org/history) in v1.8.4,
with these actions taken:

* Bump all instances of `poetry` to the new version
  * Installs in workflows
  * pre-commit hook revision
  * Dockerfiles
  * Full comparison:
    * https://github.com/python-poetry/poetry/compare/1.8.3...1.8.4
* Bump the `poetry-core` build system version in `pyproject.toml`
  * v1.9.1 is when `3.13` was added to list of available Python versions
  * This update could have waited until #488 but did now to keep current
  * Full comparison:
    * https://github.com/python-poetry/poetry-core/compare/1.8.1...1.9.1
* Update the lockfile with the new version of `poetry`

None of the changes or new features in these new versions required any
updates to the use of `poetry` in this project.