build: account for Python Semantic Release update
Python Semantic Release (PSR) released another minor version today and
it ended up having breaking changes. This time, the v9.12.0 release of
PSR included this commit, which changed another long name of the `fix`
type of angular commit as it is referred to in a parsed release element:

https://github.com/python-semantic-release/python-semantic-release/commit/09e3a4da6237740de8e9932d742b18d990e9d079

The effect would have been that, going forward, `fix` commits would not
have been picked up for inclusion. The update was accounted for in
`templates/.changes.j2`. At the same time, the section headings were
changed to title case so that all the words would be capitalized instead
of just the first word (e.g., `Bug Fixes` vs. `Bug fixes`).

These breaking changes have become a regular occurence so the dependency
was pinned to an explicit version. This will prevent more breaking
changes during the weekly automated dependency bump.

The changes in this PR were tested locally and confirmed to work for the
current configuration.