feat!: add organization support
This change adds support for organizations. It does so by adding an
optional `--org` argument to the `phylum-ci` entry point. That argument
will take precedence over any value found in the Phylum settings file.
The argument and all the logic and documentation around it allow for no
org to be specified. This is intentional since there will have to be a
transition period where some users may not be using orgs just yet.

When an org is specified, it is required that a group also be specified.
However, it is still possible to specify a group without an org and even
no group at all. The log output was updated to make it more clear which
project/org/group combo is in use.

Additional changes made include:

* Update the custom `ci` Phylum analysis extension
  * Add `--org` option and use it when calling `PhylumApi::analyze`
* Account for change in `phylum project status` json output
  * Change `repository_url` to `repositoryUrl`
  * Fixes subtle bug where existing `repo_url` entries were overwritten
* Update documentation
  * Add example usage
  * Link to https://docs.phylum.io/knowledge_base/phylum_project_files
  * Link to https://docs.phylum.io/knowledge_base/api-keys in more spots
* Refactor and format throughout
  * Create `_cmd_extender` helper function to add common CLI options
    * This also helps to keep the McCabe complexity score low for QA
  * Use the parent `OSError` exception instead of more specific one
    * This may help prevent unhandled exceptions on Windows

BREAKING CHANGE: Phylum CLI installs before v7.1.4-rc1 are no longer
supported. That release is the first one providing support for analysis
with organizations via extensions.