ci: `ctx` not available in release notes PSR env
This update changes `ctx` to `context` in the release notes environment
for Python Semantic Release (PSR). The pinned version of PSR, v9.12.0,
has a bug where `ctx` is not a valid shorthand for release notes
templates. This bug was fixed in PSR v9.12.1, but updating to a newer
version of PSR always carries a risk of breaking changes. Such an update
will be attempted at a later time.