fix: exit cleanly when no current dependencies (#503)
This change accounts for the possibility that there are no current
dependencies to analyze, even with recognized dependency files. Before
this change, the `phylum-ci` entry point would exit with an unhandled
exception and a large stack trace due to an empty analysis result being
returned from the custom `ci` extension. Now, the situation is
recognized and handled cleanly with a descriptive warning and new system
exit code of `12`.

Other changes made include:

* Remove old dead code checking for CLI exit code value of `100`
  * The analysis is performed with the custom `ci` extension now
* Use `SystemExit` in `CIPreCommit` instead of `sys.exit`
  * Exit with the current return code
* Format and refactor throughout
  * Add link to phylum project file docs in error message
  * Add initial debug log message to show `phylum` package version