fix: account for analysis started in repo sub-dirs
This change accounts for the possibility that dependency files are
specified or detected at levels lower than the root git directory. This
is the case when analysis is started in a sub-directory (e.g.,
monorepos).

When determining the base dependencies to make sure only the newly added
dependencies are considered, a temporary git worktree is created with
the common ancestor commit. That worktree is of the **entire** git repo
at that commit and therefore the relative path to dependency files
specified/detected at lower levels in the path are not correct. The
effect is an error in parsing the file due to it not existing (at the
expected path).

The fallback when this happens is to assume that all current
dependencies are new and to analyze them (e.g., the same as if the
`--all-deps` option was provided). So, in that sense, the bug is an
"over-analysis" compared to what was expected (fail secure/closed).

This change creates a path for the previous dependency file that is
relative to the git root directory instead of the current working
directory.

Other changes made include:

* Add `CIBase` class prerequisite
  * Must be operating within the context of a git repository
* Remove `Depfile` class `git` prerequisite as dead code
  * The `Depfile` class does not directly use `git`
* Format and refactor throughout