feat: avoid GitHub API calls during version check
This change updates the version processing logic to avoid GitHub API
calls when a Phylum CLI is already installed and it's version is
supported. This is especially useful for the Docker image, where there
is already a known good Phylum CLI bundled in. This change originated
with a user request where they are seeking to limit outbound network
requests.

There is also an effort made to be more intentional about the difference
between Phylum CLI versions *to be* installed vs. those that are already
installed.

Other changes made include:

* Create a new `is_installed_version_supported` predicate
  * Use it instead of scattered checks against `MIN_CLI_VER_INSTALLED`
* Update `ruff` to latest version
* Format and refactor throughout
  * Rename `is_supported_version` to `is_version_for_install_supported`
  * Extract basic version normalization code to a function