fix: `gradle` lockfile generation requires `javac`
Lockfile generation for `gradle` manifests started failing recently with
the following error message:

```
Toolchain installation '/usr/lib/jvm/java-17-openjdk-amd64' does not
provide the required capabilities: [JAVA_COMPILER]
```

This is due to
[a known regression](https://github.com/gradle/gradle/issues/30264)
introduced in Gradle v8.10. This regression is also documented in the
[upgrading guide](https://docs.gradle.org/current/userguide/upgrading_version_8.html#javacompile_tasks_may_fail_when_using_a_jre_even_if_compilation_is_not_necessary),
with a suggested fix. However, that fix would require updates in Phylum
user repositories...which is not always possible or desired. Instead,
this change updates the Docker image to include the java compiler in the
form of the system package `default-jdk-headless`. That way, the Java
JRE and JDK will match.