feat: add option to fail for incomplete analysis (#510)
This change adds a flag named `--fail-incomplete` to set a failure code
when some analysis results are unknown. This is useful in environments
where strictness is desired to prevent adding any new dependency until
it passes established policy. This is a mutually exclusive option with
`--audit`.

The name `--strict` was also considered but not used because that name
implies a "strict mode" which would be more than just the opposite of
audit mode. For instance, the ability to ignore errors with the
`--ignore-errors` flag should not be possible in a true strict mode. So,
the more specific option name was used to limit the effects of
specifying it.