build: update `poetry` to latest version (#513)
Poetry is the workflow management tool used for this project and forms
the root of all other actions taken when working with this repository.
It is also used to manage dependencies and therefore should be treated
very carefully, with updates to newer versions taken deliberately.

This PR updates `poetry` to the latest version of v1.8.5 to account for
the [changes introduced](https://python-poetry.org/history) in v1.8.5,
with these actions taken:

* Bump all instances of `poetry` to the new version
  * Installs in workflows
  * pre-commit hook revision
  * Dockerfiles
  * Full comparison:
*
[python-poetry/poetry@1.8.4...1.8.5](https://github.com/python-poetry/poetry/compare/1.8.4...1.8.5)

Attempting to update the lockfile with the new version of `poetry` did
not result in any changes, likely because the dependencies were updated
earlier today. The next weekly dependency bump will update the reference
to the version of `poetry` used to create the lockfile in `poetry.lock`.

None of the changes or new features in this new version required any
updates to the use of `poetry` in this project.