feat: automatically analyze newly created projects
This change forces analysis when a new project is created. This ensures
the project is initially populated with results in the Phylum UI, which
helps for users configuring Phylum analysis exclusively through the CI
integrations. This is doubly true when the set of dependency files do
not include a manifest, which is the other trigger for automatically
forcing analysis.

The desire expressed from users of the integrations is that setting up
the integration and running it the first time should cause the project
dependencies to be ingested without needing to set the
`--force-analysis` flag or wait until a PR with a modified dependency
file emerges. They also don't want to have to use the Phylum CLI to
populate the project(s) before scanning them with the CI integrations.